requirements.txt
Flask==3.0.3
pymongo[srv]==4.10.1
python-dotenv==1.0.1
Werkzeug==3.0.4
certifi==2024.8.30
